+++ raytracer/src/parse.h
@@ -9,18 +9,18 @@
 struct RenderOptions {
     int width = 200;
     int height = 0;
-    double aspect = 16.0 / 9.0;
+    float aspect = 16.0 / 9.0;
     vec3 lookfrom{13, 2, 3};
     vec3 lookat{0, 0, 0};
     vec3 vup{0, 1, 0};
-    double vfov = 20;
-    double aperture = 0.0;
-    double focusDist = 10.0;
+    float vfov = 20;
+    float aperture = 0.0;
+    float focusDist = 10.0;
     int samples = 100;
     int depth = 50;
     int displayWidth = 0;
     int displayHeight = 0;
-    double gamma = 2.2;
+    float gamma = 2.2;
     bool fullscreen = false;
     bool showFps = false;
     BackgroundFn background = colorBlueWhiteGradient;
